Extract schedule type-agnostic version to "Schedule"
